+++ src/Reconnections/cache.jl
@@ -7,6 +7,7 @@ using ..FindNearbySegments:
     segment_is_close,
     nearby_segments
 using ..CellLists: CellLists
+using OhMyThreads: OhMyThreads
 
 abstract type AbstractReconnectionCache end
 
@@ -112,60 +113,70 @@ function find_reconnection_pairs!(
     r_crit = T(1.5) * r_cut
     r²_crit = r_crit * r_crit
     Lhs = map(L -> L / 2, Ls)
-    function check_distance(r⃗_in)
-        r⃗ = deperiodise_separation(r⃗_in, Ls, Lhs)
-        r² = sum(abs2, r⃗)
-        r² < r²_crit
-    end
     @timeit to "set_filaments!" set_filaments!(finder, fs)  # this is needed in particular to initialise cell lists
     lck = ReentrantLock()
+    Nf = length(fs)
+    # TODO: preallocate arrays?
+    ntasks = Threads.nthreads()
+    invalidated_filaments = falses(Nf)  # this allows to add filaments at most once (ideally; in parallel that's not always the case, but that's ok for now)
+    invalidated_per_thread = [falses(Nf) for _ in 1:ntasks]
     @timeit to "find segment pairs" for (i, f) ∈ pairs(fs)
-        Threads.@threads for seg_a ∈ segments(f)
-            # Since we only compare the *midpoint* of this segment to the extrema of other
-            # segments, we add δ/2 (half the segment length) to the critical distance to take
-            # into account the case where the point of minimum distance is at the extrema of
-            # this segment (and not near the midpoint).
-            x⃗ = Filaments.midpoint(seg_a)
-            δ² = sum(abs2, f[seg_a.i + 1] - f[seg_a.i])  # ≈ squared segment length
-            d²_crit = r²_crit + δ² / 4
-            d_crit = @fastmath sqrt(d²_crit)
-            for (j, seg_b) ∈ nearby_segments(finder, x⃗)
-                # 1. Apply slightly finer filters to determine whether we keep this candidate.
-                # TODO combine these two criteria?
-                segment_is_close(seg_b, x⃗, d_crit, d²_crit, Ls, Lhs) || continue
-                keep_segment_pair(check_distance, seg_a, seg_b) || continue
-                candidate = ReconnectionCandidate(seg_a, seg_b, i, j)
-                # 2. Now check if the chosen criterion is verified
-                info = should_reconnect(crit, candidate; periods = Ls)
-                info === nothing && continue
-                # 3. Find possibly better candidates
-                # TODO: make sure there are no repetitions...
-                info, candidate = find_better_candidates(info, candidate) do other_candidate
-                    should_reconnect(crit, other_candidate; periods = Ls)
-                end
-                @assert info !== nothing
-                reconnect_info = (; candidate, info,)
-                # 4. Check if the candidate satisfies the velocity criterion
-                (; a, b, filament_idx_a, filament_idx_b,) = candidate
-                (; d⃗,) = info  # d⃗ = x⃗ - (y⃗ - p⃗)
-                if crit.use_velocity
-                    # Use velocity information to discard some reconnections
-                    @assert vs !== nothing   # checked earlier
-                    @assert eltype(vs) <: AbstractFilament  # this means it's interpolable
-                    v⃗_a = vs[filament_idx_a](a.i, info.ζx)  # assume velocity is interpolable
-                    v⃗_b = vs[filament_idx_b](b.i, info.ζy)  # assume velocity is interpolable
-                    v_d = d⃗ ⋅ (v⃗_a - v⃗_b)  # separation velocity (should be divided by |d⃗| = sqrt(d²), but we only care about the sign)
-                    if v_d > 0  # they're getting away from each other
-                        continue  # don't reconnect them
+        # Synchronise per-thread values onto invalidated_filaments.
+        for invalidated_local in invalidated_per_thread
+            @. invalidated_filaments = invalidated_filaments | invalidated_local
+        end
+        invalidated_filaments[i] && continue  # skip if this filament has already been invalidated
+        @sync for (itask, segments_task) in enumerate(OhMyThreads.index_chunks(eachindex(segments(f)); n = ntasks))
+            OhMyThreads.@spawn for i_seg in segments_task
+                invalidated_local = invalidated_per_thread[itask]
+                copyto!(invalidated_local, invalidated_filaments)  # copy most recent invalidated filaments
+                seg_a = Segment(f, i_seg)
+                # Since we only compare the *midpoint* of this segment to the extrema of other
+                # segments, we add δ/2 (half the segment length) to the critical distance to take
+                # into account the case where the point of minimum distance is at the extrema of
+                # this segment (and not near the midpoint).
+                x⃗ = Filaments.midpoint(seg_a)
+                δ² = sum(abs2, f[seg_a.i + 1] - f[seg_a.i])  # ≈ squared segment length
+                d²_crit = r²_crit + δ² / 4
+                d_crit = @fastmath sqrt(d²_crit)
+                for (j, seg_b) ∈ nearby_segments(finder, x⃗)
+                    invalidated_local[j] && continue  # skip if filament j has been invalidated
+                    # 1. Apply slightly finer filters to determine whether we keep this candidate.
+                    # TODO combine these two criteria?
+                    @inline segment_is_close(seg_b, x⃗, d_crit, d²_crit, Ls, Lhs) || continue
+                    keep_segment_pair(seg_a, seg_b) || continue
+                    candidate = ReconnectionCandidate(seg_a, seg_b, i, j)
+                    # 2. Now check if the chosen criterion is verified
+                    info = should_reconnect(crit, candidate; periods = Ls)
+                    info === nothing && continue
+                    # 3. Find possibly better candidates
+                    # TODO: make sure there are no repetitions...
+                    info, candidate = find_better_candidates(info, candidate) do other_candidate
+                        should_reconnect(crit, other_candidate; periods = Ls)
                     end
-                end
-                # 5. If the candidate passed all the tests, we add it to the reconnection list (unless it's already there)
-                lock(lck) do
-                    # Check if this candidate was already found earlier (possible due to find_better_candidates)
-                    for r ∈ to_reconnect
-                        r === reconnect_info && return  # exits `do` block (unlocks the lock without pushing candidate)
+                    (; a, b, filament_idx_a, filament_idx_b,) = candidate
+                    @assert info !== nothing
+                    reconnect_info = (; candidate, info,)
+                    # 4. Check if the candidate satisfies the velocity criterion
+                    (; d⃗,) = info  # d⃗ = x⃗ - (y⃗ - p⃗)
+                    if crit.use_velocity
+                        # Use velocity information to discard some reconnections
+                        @assert vs !== nothing   # checked earlier
+                        @assert eltype(vs) <: AbstractFilament  # this means it's interpolable
+                        v⃗_a = vs[filament_idx_a](a.i, info.ζx)  # assume velocity is interpolable
+                        v⃗_b = vs[filament_idx_b](b.i, info.ζy)  # assume velocity is interpolable
+                        v_d = d⃗ ⋅ (v⃗_a - v⃗_b)  # separation velocity (should be divided by |d⃗| = sqrt(d²), but we only care about the sign)
+                        if v_d > 0  # they're getting away from each other
+                            continue  # don't reconnect them
+                        end
                     end
-                    push!(to_reconnect, reconnect_info)
+                    # 5. If the candidate passed all the tests, we add it to the reconnection list (unless it's already there)
+                    invalidated_local[filament_idx_a] = true
+                    invalidated_local[filament_idx_b] = true
+                    lock(lck) do
+                        @inline
+                        push!(to_reconnect, reconnect_info)
+                    end
                 end
             end
         end
@@ -211,19 +222,14 @@ end
 
 # This is to make sure we don't reconnect nearly neighbouring segments belonging to the same
 # filament.
-function keep_segment_pair(check_distance::F, a::Segment, b::Segment) where {F <: Function}
+function keep_segment_pair(a::Segment, b::Segment)
     f, i = a.f, a.i
     g, j = b.f, b.i
     if f === g
-        i === j && return false  # same segment
         dist_max = 2  # disallow reconnection between segment i and segment j = i ± {0, 1, 2}
         N = length(f)
-        Nh = N >> 1
-        i′, j′ = ifelse(i < j, (i, j), (j, i))
-        while j′ - i′ > Nh
-            i′, j′ = j′ - N, i′
-        end
-        if j′ - i′ ≤ dist_max
+        dist = abs(i - j)
+        if dist ≤ dist_max || (N - dist) ≤ dist_max  # note: N - dist is always ≥ 0
             return false
         end
     end