+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.28.6] - 2025-07-04
+
 ### Performance
 
 - Various major performance improvements to reconnections.